backend/requirements.txt
fastapi==0.115.14
uvicorn[standard]==0.35.0
pydantic==2.11.7
python-dotenv==1.0.1
requests==2.32.4
